+++ src/lib/connection/XMPPServiceHandler/channelEventHandler.ts
@@ -310,7 +310,7 @@ class ChannelEventHandler extends GenericHandler {
                     // // Treated in conversation handler that.logger.log("error", LOG_ID + "(onErrorMessageReceived) The 'to' of the message can not received the message");
                 } else {
                     that.logger.log("internalerror", LOG_ID + "(onErrorMessageReceived) something goes wrong...", msg, stanza);
-                    that.eventEmitter.emit("rainbow_onerror", msg);
+                    that.eventEmitter.emit("evt_internal_xmpperror", msg);
                 }
             } catch (err) {
                 that.logger.log("error", LOG_ID + "(onErrorMessageReceived) CATCH Error !!! ");