+++ src/lib/connection/XMPPServiceHandler/conversationEventHandler.ts
@@ -1024,7 +1024,7 @@ class ConversationEventHandler extends GenericHandler {
                 } else {
                     that.logger.log("error", LOG_ID + "(onErrorMessageReceived) something goes wrong...");
                     that.logger.log("internalerror", LOG_ID + "(onErrorMessageReceived) something goes wrong... : ", msg, util.inspect(stanza));
-                    that.eventEmitter.emit("rainbow_onerror", msg);
+                    that.eventEmitter.emit("evt_internal_xmpperror", msg);
                 }
             } catch (err) {
                 that.logger.log("error", LOG_ID + "(onErrorMessageReceived) CATCH Error !!! ");