+++ CHANGELOG.md
@@ -8,7 +8,12 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 
 ## [1.66.0] - 2020-01-XX
 -   Add correlatorData et GlobaleCallId properties in Call type of phone calls : RQRAINB-2773, RQRAINB-2784, RQRAINB-2784, RQRAINB-2789, RQRAINB-2793, RQRAINB-2793, RQRAINB-2799
--   Fix method ChannelsService::createItem when parameter "type" is setted. 
+-   Fix method ChannelsService::createItem when parameter "type" is setted.
+-   Split Xmmpp error event treatment in 3 possibilities:
+    * Errors which need a reconnection 
+    * Errors which need to only raise an event to inform up layer. => Add an event `rainbow_onxmpperror` to inform about issue. 
+    * Errors which are fatal errors and then need to stop the SDK. => Already existing events `rainbow_onerror` + `rainbow_onstop`.
+-   work done on private method BubblesServices::joinConference (Not finish, so not available). 
 
 ## [1.65.2] - 2020-01-10
 -   remove an unwanted log